Added Latin Mass Propers support to Lectionary
With this update Latin Mass Propers for the day can be displayed, with 12 updates per day for each part of the Mass.

+++ libraries/RomanCalendar/RomFont.h
@@ -11,8 +11,11 @@ typedef struct
 	const uint8_t widthBits;				// width, in bits (or pixels), of the character
 	const uint8_t heightBits;				// height, in bits (or pixels), of the character
 	const uint32_t offset;					// offset of the character's bitmap, in bytes, into the the FONT_INFO's data array
-	const double advanceWidth;				// the number of pixels to advance before placing the next character
-	const double advanceHeight;				// the number of pixels to advance vertically before placing the next character
+	//PLL-30-07-2020 changed these to floats from doubles - saves 3% flash
+	//const double advanceWidth;				// the number of pixels to advance before placing the next character
+	//const double advanceHeight;				// the number of pixels to advance vertically before placing the next character
+	const float advanceWidth;				// the number of pixels to advance before placing the next character
+	const float advanceHeight;				// the number of pixels to advance vertically before placing the next character
 } FONT_CHAR_INFO;	
 
 typedef struct
